Added most of the remaining battle messages. Made necessary changes to related structures. Battle info is now much more colorful.

+++ src/com/pokebros/android/pokemononline/ColorEnums.java
@@ -60,7 +60,7 @@ public class ColorEnums {
 		
 	public static class StatusColor {
 		private final String color;
-		public StatusColor(byte status) {
+		public StatusColor(int status) {
 			switch (Status.values()[status]) {
 			case Koed: color = "#171ba>"; break;
 			case Fine: color = TypeColor.Normal.toString(); break;
